+++ spec/unit/provider/group/dscl_spec.rb
@@ -240,6 +240,7 @@ describe Chef::Provider::Group::Dscl do
       @provider.load_current_resource
       @provider.define_resource_requirements
     end
+
     it "raises an error if the required binary /usr/bin/dscl doesn't exist" do
       File.should_receive(:exists?).with("/usr/bin/dscl").and_return(false)
 
@@ -251,7 +252,7 @@ describe Chef::Provider::Group::Dscl do
       lambda { @provider.process_resource_requirements }.should_not raise_error
     end
   end
-
+ 
   describe "when creating the group" do
     it "creates the group, password field, gid, and sets group membership" do
       @provider.should_receive(:set_gid).and_return(true)
@@ -292,5 +293,41 @@ describe Chef::Provider::Group::Dscl do
       @provider.should_receive(:safe_dscl).with("delete /Groups/aj").and_return(true)
       @provider.remove_group
     end
+  end
+end
+
+describe 'Test DSCL loading' do
+  before do
+    @node = Chef::Node.new
+    @events = Chef::EventDispatch::Dispatcher.new
+    @run_context = Chef::RunContext.new(@node, {}, @events)
+    @new_resource = Chef::Resource::Group.new("aj")
+    @provider = Chef::Provider::Group::Dscl.new(@new_resource, @run_context)
+    @output = <<-EOF
+AppleMetaNodeLocation: /Local/Default
+Comment:
+ Test Group
+GeneratedUID: AAAAAAAA-AAAA-AAAA-AAAA-AAAAAAAAAAAA
+NestedGroups: AAAAAAAA-AAAA-AAAA-AAAA-AAAAAAAAAAAB
+Password: *
+PrimaryGroupID: 999
+RealName:
+ TestGroup
+RecordName: com.apple.aj
+RecordType: dsRecTypeStandard:Groups
+GroupMembership: waka bar
+EOF
+    @provider.stub(:safe_dscl).with("read /Groups/aj").and_return(@output)
+    @current_resource = @provider.load_current_resource
+
+  end
+
+  it 'should parse gid properly' do
+    File.stub(:exists?).and_return(true)
+    @current_resource.gid.should eq("999")
+  end
+  it 'should parse members properly' do
+    File.stub(:exists?).and_return(true)
+    @current_resource.members.should eq(['waka', 'bar'])
   end
 end